Base Player Creation
Base players have been created. There is a little bit of jank with them, but that can be looked into later. Obviously, a few things are likely to change as we get more refined art. The player prefab does not have a camera enabled by default. If you are placing a human player, you will need to check that on. In all cases, you will need to add a player controller for the player to work. A few controllers are provided in this commit.